GEOS-7960: Factoring out GeoServerNodeData class from GeoServerNodeInfo.
(cherry picked from commit 0009f76dc02c64a43455d801224b4acb969946f2)
