+++ src/web/core/src/main/java/org/geoserver/web/GeoServerNodeInfo.java
@@ -6,6 +6,7 @@
 package org.geoserver.web;
 
 import org.apache.wicket.markup.html.WebMarkupContainer;
+import org.geoserver.GeoServerNodeData;
 
 /**
  * Provides identification and styling for the GeoServer id element showing right below the
@@ -22,7 +23,16 @@ interface GeoServerNodeInfo {
      * 
      *
      */
-    String getId();
+    default String getId() {
+        return getData().getId();
+    }
+
+    /**
+     * Returns the data object containing the node info. 
+     */
+    default GeoServerNodeData getData() {
+        return GeoServerNodeData.createFromEnvironment();
+    }
 
     /**
      * Allows customization of the label container, in particular, its style and visibility, but
